requirements.txt
quantifin
boto3
jupyter
pyyaml
jinja2
weasyprint
